add initial commit
got implementation plan using:

```
Look carefully at instructions.txt and in the resources/ directory. In particular, note the resources/sign/ and resources/id/ pair of example applications

Expand instructions.txt into a detailed implementation plan. The implementor will be starting from a copy of the id/ app as a template

Note in particular that bindings for the UI will be generated when the app is built with `kit build --hyperapp`. As such, first design and implement the backend; from the backend, the interface will be generated; finally design and implement the frontend to consume the interface. Subsequent changes to the interface must follow this pattern as well: start in backend, generated interface, finish in frontend

`Address` is the fully-qualified address of a process, NOT a node identity. It contains within it the node identity. An Address looks like

user.os@process:package:publisher.os

where `user.os` is the node running the process, `process` is the process name`, `package` is the package name, and `publisher.os` is the publisher of the package
```

```
the WIT API is entirely machine generated based on the function signatures defined in the hyperapp `lib.rs`. Types should be defined there and they will be machine generated into the WIT API: do not write the WIT API
```

```
- `#[on_init]` should be `#[init]`
- #[http] must be specified either as #[http] for all methods and paths or #[http(method = "POST")] for POSTs on all paths or #[http(method = "POST", path = "/foo")] for specifically POSTs on path `/foo`
- #[ws] indicates websocket handler
```